requirements.txt
pandas
discord.py
mariadb